pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thor"
version = "0.1.0"
description = "Travel Order Resolver - NLP system for extracting train travel orders"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "THOR Team"}
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
stt = [
    "openai-whisper>=20231117",
    "vosk>=0.3.45",
    "google-cloud-speech>=2.19.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "jiwer>=3.0.0",
]
nlp = [
    "spacy>=3.7.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "datasets>=2.14.0",
    "seqeval>=1.2.2",
]
pathfinding = [
    "neo4j>=5.0.0",
    "networkx>=3.0",
    "scipy>=1.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
tts = [
    "gtts>=2.3.0",
    "pyttsx3>=2.90",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
thor-stt = "src.cli.stt:main"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

